Fix blocking reads for valkey-benchmark over RDMA.
There are blocking reads when the data size is larger than 16KB.
In valkey-benchmark.c, valkeyBufferRead reads 16KB at most, and
go to sleep if full reply message has not been received. For TCP,
the event loop will be waked up immediately if there is
data left in the buffer. However, as the ready-to-read event is
edge-triggered in RDMA, it will be blocked forever. This patch
makes valkeyBufferRead read all the available data in the buffer.

Change-Id: I81093f1d019bc00175448aae3a4e8b6dca4ffd39
Signed-off-by: ruihong <ruihong@google.com>